requirements.txt
jsonschema>=4.21.0
PyYAML>=6.0.1
pytest>=7.0.0
requests
ruff>=0.6.0
